Cambio tipo variable recursos a int

+++ Assets/Scripts/App/Events/PlayerGetResourceEvent.cs
@@ -5,7 +5,7 @@ namespace App.Events
     [CreateAssetMenu(fileName = "PlayerGetMaterialEvent", menuName = "Events/Resources/PlayerGetMaterialEvent")]
     public class PlayerGetResourceEvent : GameEventScriptable
     {
-        public float Quantity;
+        public int Quantity;
         public RetrievableResourceType Type;
     }
 }
