+++ Assets/Scripts/App/Models/IResourcesModel.cs
@@ -2,8 +2,8 @@
 {
     public interface IResourcesModel
     {
-        float Gold { get; set; }
-        float Metal { get; set; }
-        void AddResources(RetrievableResourceType type, float quantity);
+        int Gold { get; set; }
+        int Metal { get; set; }
+        void AddResources(RetrievableResourceType type, int quantity);
     }
 }
